backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
dnspython==2.4.2
